refactor: combine email check with error checks and extend return structure of userData
