+++ src/components/Register/RegisterUser.ts
@@ -1,9 +1,10 @@
-import { RegisterFormFields } from './types';
-import mapFormData from './FormUserData';
 import { Customer } from '@commercetools/platform-sdk';
 import { createAnonymousClient } from '@/services/commercetools/client/createAnonymousClient';
 import { toast } from 'sonner';
-import handleRegError from './RigisterErrorHandler';
+
+import { RegisterFormFields } from './types';
+import mapFormData from './formUserData';
+import { handleRegError } from './registerUtils';
 
 const registerUser = async (userData: RegisterFormFields): Promise<Customer | undefined> => {
   const apiRoot = createAnonymousClient();
@@ -11,7 +12,6 @@ const registerUser = async (userData: RegisterFormFields): Promise<Customer | un
   const userDraft = {
     ...mapFormData(userData)
   };
-
   try {
     await apiRoot.customers().post({ body: userDraft }).execute();
